move extrinsics utilities to extrinsics module
