+++ crates/sp-domains/src/lib.rs
@@ -18,6 +18,7 @@
 #![cfg_attr(not(feature = "std"), no_std)]
 
 pub mod bundle_producer_election;
+pub mod extrinsics;
 pub mod fraud_proof;
 pub mod merkle_tree;
 pub mod storage;
