+++ crates/sp-domains/src/verification.rs
@@ -1,22 +1,14 @@
-use crate::{ExecutionReceipt, DOMAIN_EXTRINSICS_SHUFFLING_SEED_SUBJECT};
-use domain_runtime_primitives::opaque::AccountId;
+use crate::ExecutionReceipt;
 use frame_support::PalletError;
 use hash_db::Hasher;
 use parity_scale_codec::{Compact, Decode, Encode};
-use rand::seq::SliceRandom;
-use rand::SeedableRng;
-use rand_chacha::ChaCha8Rng;
 use scale_info::TypeInfo;
 use sp_core::storage::StorageKey;
 use sp_runtime::traits::{Block, NumberFor};
-use sp_state_machine::trace;
-use sp_std::collections::btree_map::BTreeMap;
-use sp_std::collections::vec_deque::VecDeque;
 use sp_std::fmt::Debug;
 use sp_std::marker::PhantomData;
 use sp_std::vec::Vec;
 use sp_trie::{read_trie_value, LayoutV1, StorageProof};
-use subspace_core_primitives::Randomness;
 
 /// Verification error.
 #[derive(Debug, PartialEq, Eq, Encode, Decode, PalletError, TypeInfo)]
@@ -125,80 +117,4 @@ where
     }
 
     Ok(())
-}
-
-pub fn extrinsics_shuffling_seed<Hashing>(block_randomness: Randomness) -> Hashing::Out
-where
-    Hashing: Hasher,
-{
-    let mut subject = DOMAIN_EXTRINSICS_SHUFFLING_SEED_SUBJECT.to_vec();
-    subject.extend_from_slice(block_randomness.as_ref());
-    Hashing::hash(&subject)
-}
-
-pub fn deduplicate_and_shuffle_extrinsics<Extrinsic>(
-    mut extrinsics: Vec<(Option<AccountId>, Extrinsic)>,
-    shuffling_seed: Randomness,
-) -> VecDeque<Extrinsic>
-where
-    Extrinsic: Debug + PartialEq + Clone,
-{
-    let mut seen = Vec::new();
-    extrinsics.retain(|(_, uxt)| match seen.contains(uxt) {
-        true => {
-            trace!(extrinsic = ?uxt, "Duplicated extrinsic");
-            false
-        }
-        false => {
-            seen.push(uxt.clone());
-            true
-        }
-    });
-    drop(seen);
-    trace!(?extrinsics, "Origin deduplicated extrinsics");
-    shuffle_extrinsics::<Extrinsic, AccountId>(extrinsics, shuffling_seed)
-}
-
-/// Shuffles the extrinsics in a deterministic way.
-///
-/// The extrinsics are grouped by the signer. The extrinsics without a signer, i.e., unsigned
-/// extrinsics, are considered as a special group. The items in different groups are cross shuffled,
-/// while the order of items inside the same group is still maintained.
-pub fn shuffle_extrinsics<Extrinsic: Debug, AccountId: Ord + Clone>(
-    extrinsics: Vec<(Option<AccountId>, Extrinsic)>,
-    shuffling_seed: Randomness,
-) -> VecDeque<Extrinsic> {
-    let mut rng = ChaCha8Rng::from_seed(*shuffling_seed);
-
-    let mut positions = extrinsics
-        .iter()
-        .map(|(maybe_signer, _)| maybe_signer)
-        .cloned()
-        .collect::<Vec<_>>();
-
-    // Shuffles the positions using Fisher–Yates algorithm.
-    positions.shuffle(&mut rng);
-
-    let mut grouped_extrinsics: BTreeMap<Option<AccountId>, VecDeque<_>> = extrinsics
-        .into_iter()
-        .fold(BTreeMap::new(), |mut groups, (maybe_signer, tx)| {
-            groups.entry(maybe_signer).or_default().push_back(tx);
-            groups
-        });
-
-    // The relative ordering for the items in the same group does not change.
-    let shuffled_extrinsics = positions
-        .into_iter()
-        .map(|maybe_signer| {
-            grouped_extrinsics
-                .get_mut(&maybe_signer)
-                .expect("Extrinsics are grouped correctly; qed")
-                .pop_front()
-                .expect("Extrinsic definitely exists as it's correctly grouped above; qed")
-        })
-        .collect::<VecDeque<_>>();
-
-    trace!(?shuffled_extrinsics, "Shuffled extrinsics");
-
-    shuffled_extrinsics
 }
